requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
Django==4.0.4
django-cleanup==6.0.0
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
feedparser==6.0.10
gunicorn==20.1.0
Markdown==3.3.7
mysqlclient==2.1.0
Pillow==9.1.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
